requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
asttokens==2.2.1
    # via stack-data
backcall==0.2.0
    # via ipython
black==23.3.0
    # via -r requirements.in
build==0.10.0
    # via pip-tools
certifi==2023.5.7
    # via sentry-sdk
click==8.1.3
    # via
    #   black
    #   pip-tools
coverage[toml]==7.2.7
    # via pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django==4.2.2
    # via
    #   -r requirements.in
    #   django-configurations
    #   django-debug-toolbar
    #   django-stubs
    #   django-stubs-ext
    #   model-bakery
django-configurations==2.4.1
    # via -r requirements.in
django-debug-toolbar==4.1.0
    # via -r requirements.in
django-stubs[compatible-mypy]==4.2.1
    # via -r requirements.in
django-stubs-ext==4.2.1
    # via django-stubs
executing==1.2.0
    # via stack-data
gunicorn==20.1.0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r requirements.in
ipython==8.14.0
    # via
    #   -r requirements.in
    #   ipdb
isort==5.12.0
    # via -r requirements.in
jedi==0.18.2
    # via ipython
matplotlib-inline==0.1.6
    # via ipython
model-bakery==1.12.0
    # via -r requirements.in
mypy==1.3.0
    # via
    #   -r requirements.in
    #   django-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via -r requirements.in
pip-tools==6.13.0
    # via -r requirements.in
platformdirs==3.7.0
    # via black
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.38
    # via ipython
psycopg2-binary==2.9.6
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.15.1
    # via ipython
pyproject-hooks==1.0.0
    # via build
pytest==7.3.2
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.5.2
    # via -r requirements.in
ruff==0.0.274
    # via -r requirements.in
sentry-sdk==1.26.0
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
types-pytz==2023.3.0.0
    # via django-stubs
types-pyyaml==6.0.12.10
    # via django-stubs
typing-extensions==4.6.3
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
urllib3==2.0.3
    # via sentry-sdk
wcwidth==0.2.6
    # via prompt-toolkit
wheel==0.40.0
    # via pip-tools
whitenoise==6.5.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
